Added AnyUserProvider. Added PHPUnit tests.
