+++ .travis.yml
@@ -1,0 +1,11 @@
+language: php
+
+php:
+    - 5.6
+
+before_script:
+    - curl -s https://getcomposer.org/installer | $(which php)
+    - $(which php) composer.phar install
+
+script:
+    - vendor/bin/phpunit
