+++ composer.json
@@ -16,6 +16,9 @@
 		"symfony/security": "^3.0",
 		"symfony/ldap": "^3.0"
 	},
+	"require-dev": {
+		"phpunit/phpunit": "5.7"
+	},
 	"suggest": {
 		"mlukman/securilex-doctrine": "For loading users from database using Doctrine ORM"
 	},
